refactor: remove sol-specific code from re-usable component
